Refactor project dir
