+++ test_files/connection/response/getConnectionGraph_200.json
